backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
PyPDF2==3.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai==1.3.7
httpx==0.25.2
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
